+++ ClassicLootManager/Modules/Points/AutoAwardManager/EncounterIDs.lua
@@ -228,6 +228,9 @@ CLM.EncounterHasHardMode = CLM.UTILS.Set({
 })
 
 CLM.EncounterIDs = {}
+
+if not CLM.WoWSoD then return end
+
 CLM.EncounterIDs.Classic = {
     {
         name = CLM.L["Onyxia's Lair"],